+++ jscomp/bin/bsb.ml
@@ -9390,7 +9390,7 @@ let (|?)  m (key, cb) =
   m  |> Ext_json.test key cb 
 
 
-let warning_unused_file : _ format = "@{<warning>IGNORED@}: file %s under %s is ignored due to that it is not a valid module name@."
+let warning_unused_file : _ format = "@{<warning>IGNORED@}: file %s under %s is ignored because it can't be turned into a valid module name. The build system transforms a file name into a module name by upper-casing the first letter@."
 
 type cxt = {
   no_dev : bool ;
@@ -12490,7 +12490,11 @@ let bsb_main_flags : (string * Arg.spec * string) list=
     "-query", Arg.String (fun s -> Bsb_query.query ~cwd ~bsc_dir s ),
     " (internal)Query metadata about the build";
     "-themes", Arg.Unit Bsb_init.list_themes,
-    " List all available themes"
+    " List all available themes";
+    "-where",
+       Arg.Unit (fun _ -> 
+        print_endline (Filename.dirname Sys.executable_name)),
+    " Show where bsb.exe is located"
   ]